requirements.txt
flet
requests
openai-whisper
numpy
